+++ CHANGELOG.md
@@ -9,7 +9,11 @@ MINOR
 ### StripeCore
 * [Added] Added `additionalHeaders` support to the `STPAPIClient` GET, POST, and DELETE APIs exposed through the `STP` SPI.
 
+### Payments
+* [Added] Added billing address and phone support to `STPCollectBankAccountParams` for US bank accounts.
+
 ### PaymentSheet
+* [Fixed] Fixed US bank account collection omitting configured billing addresses and phone numbers.
 * [Fixed] Fixed card funding warnings not appearing after scanning a card.
 
 ## 26.11.0 2026-09-14